// src/types/api/userResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0
 */

export interface UserResponseDto {
  /** The unique identifier of the user */
  id: number;
  /** The email address of the user */
  email: string;
  /**
   * The first name of the user
   * @nullable
   */
  firstName?: string | null;
  /**
   * The last name of the user
   * @nullable
   */
  lastName?: string | null;
  /** Whether the user email is verified */
  isVerified: boolean;
  /** The date the user was created */
  createdAt: string;
  /** The date the user was last updated */
  updatedAt: string;
}